[kernel][ipc] Abandon shared page payload
"Shared page" is introduced to support shared memory style communication but
it turns out it makes memmgr difficult to track whether a page is being used
or not on process destruction. Consider the following example:

       1. shared a page            2.share the page
memmgr -----------------> server A -----------------> server B
  ^                                                     |
  |                                                     |
  +-----------------------------------------------------+
      3. B terminates and kernel notifies to the memmgr

Memmgr, A and B shares a same page P. On termination of B, memmgr is
notified (not yet implemented though) by the kernel but it cannot determine
whether the page P is shared with A and B or not.

If A shares it with B, memmgr should not free the page because it is still
being used by A. If not, memmgr should free the page because it is no longer used.

Using "move page" does not incur the problem. Let's stick with it for now.