kernel: Add fine-grained IPC capabilty (CAP_IPC_WITH)
While this mechanism is coarse-grained compared to other capability-based
systems like seL4, it's sufficient for us.

As you can see, it adds a restriction that TASKS_MAX must be less than
or equals to 60. That said, I believe it's large enough for small systems.

This commit also introduces NOTIFY_CLOSED notifications. It enables
servers to free unclosed handles when a client exits (not yet supported in
std, btw).