Remove kernel heap (closes #14)
Currently, Resea Kernel uses its internal static-sized heap for allocating kernel
stacks and page tables. However, as described in microkernel papers authored by
L4 folks, handling page faults in (bulk) IPC messes up its implementation and
it's not good for the separation of mechanism and policy.

This commit removes the kernel heap by:

- Add "map" system call to allow manipulating the page table from the userland.
- Allocate kernel's internal task data structures statically.
- Implement "bulkcopy" mechanism in bootstrap server.

The disadvantage of this approach is that it degrades the bulk IPC performance
since it requires additional IPC with bootstrap server. That said, since the
bootstrap server has authority to manipulate page tables, I believe we can
implement zero-copy bulk IPCs in the future for performance without modifying
the kernel.